Initial commit
Currently the script can be used to rename tv series with their
respective episode number and names.

v0.0.1 reads files in an increasing order of naming, and applies the
renames them based on the data from server. If number of episodes in
the system doesn't match the one provided by server then a message is
shown saying Episode count differs, can't continue with renaming.