Add checks for video and subtitle files
A check is added if a file is a video file or a subtitle or whether
it belongs to trash. There are global variable lists which decides
the video and subtitle formats supported now. The file names are now
changed with more precision instead of renaming every file in the list.